+++ src/main/java/com/project/doubleshop/web/config/security/Role.java
@@ -1,4 +1,4 @@
-package com.project.doubleshop.web.security;
+package com.project.doubleshop.web.config.security;
 
 public enum Role {
