+++ src/main/java/com/project/doubleshop/web/config/security/SimpleAuthenticationToken.java
@@ -1,9 +1,11 @@
-package com.project.doubleshop.web.security;
+package com.project.doubleshop.web.config.security;
 
 import java.util.Collection;
 
 import org.springframework.security.authentication.AbstractAuthenticationToken;
 import org.springframework.security.core.GrantedAuthority;
+
+import com.project.doubleshop.web.member.dto.AuthenticationRequest;
 
 public class SimpleAuthenticationToken extends AbstractAuthenticationToken {
