+++ src/main/java/com/project/doubleshop/web/config/security/SimpleToken.java
@@ -1,4 +1,4 @@
-package com.project.doubleshop.web.security;
+package com.project.doubleshop.web.config.security;
 
 import java.util.Arrays;
 import java.util.Date;
