+++ src/main/java/com/project/doubleshop/web/config/security/SimpleTokenConfigurer.java
@@ -1,7 +1,6 @@
-package com.project.doubleshop.web.security;
+package com.project.doubleshop.web.config.security;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Component;
 
 @Component
