+++ src/main/java/com/project/doubleshop/web/config/security/UriBasedVoter.java
@@ -1,4 +1,4 @@
-package com.project.doubleshop.web.security;
+package com.project.doubleshop.web.config.security;
 
 import static org.springframework.util.ClassUtils.*;
