+++ src/main/java/com/project/doubleshop/web/member/dto/AuthenticationResult.java
@@ -1,6 +1,6 @@
-package com.project.doubleshop.web.security;
+package com.project.doubleshop.web.member.dto;
 
-import com.project.doubleshop.domain.member.entity.Member;
+import com.project.doubleshop.web.config.security.SimpleToken;
 
 import lombok.Getter;
